requirements.txt
python>=3.8.0
numpy>=1.19.0
pandas>=1.1.3
pillow>=7.2.0
pytorch==1.7.0
torchvision>=0.8.1
scipy>=1.5.2
tqdm>=4.53.0
pytz>=2020.4
outdated>=0.2.0
ogb>=1.2.3
wilds>=1.0.0
transformers>=4.1.1
